Upgraded checkout app to Materialize
changes were made to the way Stripe charges are handled to avoid an error being thrown on cards.
Remaining is catching an invisible Stripe bug where the javascript errors out and payment just seems to not do anything.
Form changes are merely textual for the labels returned to the form.